+++ circuits/core/manager.py
@@ -2,6 +2,7 @@
 This module defines the Manager class.
 """
 import atexit
+import types
 from collections import deque
 from heapq import heappop, heappush
 from inspect import isfunction
@@ -17,7 +18,6 @@ from traceback import format_exc
 from types import GeneratorType
 from uuid import uuid4 as uuid
 
-from ..six import Iterator, create_bound_method, next
 from ..tools import tryimport
 from .events import Event, exception, generate_events, signal, started, stopped
 from .handlers import handler
@@ -60,7 +60,7 @@ class ExceptionWrapper:
         return self.exception
 
 
-class Sleep(Iterator):
+class Sleep:
 
     def __init__(self, seconds):
         self._task = None
@@ -389,7 +389,7 @@ class Manager:
         return handlers
 
     def addHandler(self, f):
-        method = create_bound_method(f, self) if isfunction(f) else f
+        method = types.MethodType(f, self) if isfunction(f) else f
 
         setattr(self, method.__name__, method)