+++ circuits/net/utils.py
@@ -1,7 +1,5 @@
 """Utilities"""
 
-from circuits.six import iterbytes
-
 
 def is_ssl_handshake(buf):
     """Detect an SSLv2 or SSLv3 handshake"""
@@ -12,6 +10,6 @@ def is_ssl_handshake(buf):
         return True
 
     # SSLv2
-    v = list(iterbytes(buf[:2])) + [0x00, 0x00]
+    v = list(iter(buf[:2])) + [0x00, 0x00]
     if (v[0] & 0x80 == 0x80) and ((v[0] & 0x7F) << 8 | v[1]) > 9:
         return True