+++ circuits/tools/__init__.py
@@ -7,8 +7,6 @@ import inspect as _inspect
 from functools import wraps
 from warnings import warn, warn_explicit
 
-from circuits.six import _func_code
-
 
 def tryimport(modules, obj=None, message=None):
     modules = (modules,) if isinstance(modules, str) else modules
@@ -204,8 +202,8 @@ def deprecated(f):
         warn_explicit(
             f"Call to deprecated function {f.__name__}",
             category=DeprecationWarning,
-            filename=getattr(f, _func_code).co_filename,
-            lineno=getattr(f, _func_code).co_firstlineno + 1,
+            filename=getattr(f, "__code__").co_filename,
+            lineno=getattr(f, "__code__").co_firstlineno + 1,
         )
         return f(*args, **kwargs)