+++ circuits/web/http.py
@@ -5,12 +5,11 @@ or commonly known as HTTP.
 """
 from io import BytesIO
 from socket import socket
+from urllib.parse import quote
 
 from circuits.core import BaseComponent, Value, handler
 from circuits.net.events import close, write
 from circuits.net.utils import is_ssl_handshake
-from circuits.six import text_type
-from circuits.six.moves.urllib_parse import quote
 
 from . import wrappers
 from .constants import SERVER_PROTOCOL, SERVER_VERSION
@@ -94,7 +93,7 @@ class HTTP(BaseComponent):
         sock = res.request.sock
 
         if data is not None:
-            if isinstance(data, text_type):
+            if isinstance(data, str):
                 data = data.encode(self._encoding)
 
             if res.chunked:
@@ -163,7 +162,7 @@ class HTTP(BaseComponent):
 
             if isinstance(res.body, bytes):
                 body = res.body
-            elif isinstance(res.body, text_type):
+            elif isinstance(res.body, str):
                 body = res.body.encode(self._encoding)
             else:
                 parts = (s if isinstance(s, bytes) else s.encode(self._encoding) for s in res.body if s is not None)