+++ circuits/web/processors.py
@@ -1,8 +1,6 @@
 import re
 from cgi import parse_header
 
-from circuits.six import PY3
-
 from .headers import HeaderElement
 from .parsers import MultipartParser, QueryStringParser
 
@@ -37,9 +35,7 @@ def process_multipart(request, params):
 
 def process_urlencoded(request, params, encoding="utf-8"):
     params.update(QueryStringParser(request.qs).result)
-    body = request.body.getvalue()
-    if PY3:
-        body = body.decode(encoding)
+    body = request.body.getvalue().decode(encoding)
     result = QueryStringParser(body).result
     for key, value in result.items():
         params[_decode_value(key, encoding)] = _decode_value(value, encoding)